Created a file for each message handler
